Big overhaul - fix memeory leaks, more standard compliant
Big cleanup, removed the old hashtable code, it wasn't
accessible in this package anyway. If you still want that
api, use the hashtable npm.

Fixed some memory leaks, I was using persistent values too
much. I also brought it more in line with the spec, you can
now pass in an iterator (from a different Map or anything
that has a .next(), .done, .key and .value properties) to
the constructor. Also different key objects that are deeply
equal are no longer considered the same. I think this sucks,
but it's what the spec says. As a consequence, the hashing
function should be faster and I'm bumping the major version,
it may not be compatible with code already out there.